+++ oci/core_instance_resource.go
@@ -283,6 +283,10 @@ func InstanceResource() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
+						"is_consistent_volume_naming_enabled": {
+							Type:     schema.TypeBool,
+							Computed: true,
+						},
 						"is_pv_encryption_in_transit_enabled": {
 							Type:     schema.TypeBool,
 							Computed: true,