+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for `retry_duration_seconds` option to configure length of retry in the face of HTTP 429 and 500 errors
 - Support for custom header insertion, extension, and removal for Load Balancer listener resource
+- Support for consistent volume names in the Block Volume attachments
 
 ### Fixed
 - Retried SDK calls are now jittered to avoid herding of retry requests in high parallelism scenarios